move storage errors
