+++ storage/pebble/blocks.go
@@ -12,8 +12,8 @@ import (
 	"github.com/onflow/go-ethereum/common"
 
 	"github.com/onflow/flow-evm-gateway/models"
+	errs "github.com/onflow/flow-evm-gateway/models/errors"
 	"github.com/onflow/flow-evm-gateway/storage"
-	errs "github.com/onflow/flow-evm-gateway/storage/errors"
 )
 
 var _ storage.BlockIndexer = &Blocks{}
