+++ storage/pebble/receipts.go
@@ -14,8 +14,8 @@ import (
 	"github.com/onflow/go-ethereum/rlp"
 
 	"github.com/onflow/flow-evm-gateway/models"
+	errs "github.com/onflow/flow-evm-gateway/models/errors"
 	"github.com/onflow/flow-evm-gateway/storage"
-	errs "github.com/onflow/flow-evm-gateway/storage/errors"
 )
 
 var _ storage.ReceiptIndexer = &Receipts{}
